+++ Space_Storm/Projectile.hpp
@@ -5,17 +5,17 @@
 class Projectile : public Entity 
 {
 private:
-    unsigned int m_damage;
-    int m_velocity;
+    byte m_damage;
+    byte m_velocity;
     bool m_isActive;
 public:
     Projectile();
 
-    void setDamage(unsigned int damage);
-    int getDamage() const;
+    void setDamage(byte damage);
+    byte getDamage() const;
 
-    void setVelocity(int velocity);
-    int getVelocity() const;
+    void setVelocity(byte velocity);
+    byte getVelocity() const;
     
     void setIsActive(bool isActive);
     bool isActive() const;
