+++ Space_Storm/Vector2.cpp
@@ -8,7 +8,7 @@ Vector2::Vector2()
 
 }
 
-Vector2::Vector2(int vecX, int vecY)
+Vector2::Vector2(byte vecX, byte vecY)
 :x{vecX}
 ,y{vecY}
 {
